refactor: extract Form_Records_Filters from Form_Submissions
Move the Pro list-table filter UI (form/post dropdowns, parse_query handler,
locked upsell) into a dedicated Form_Records_Filters class registering its own
restrict_manage_posts/parse_query hooks. Add a characterization test for the
Pro filter query via the parse_query hook.
